changing some things hackery

+++ Radiant/fileHandler.js
@@ -1,5 +1,6 @@
 require('dotenv').config();
 
+const { spawn } = require('child_process');
 const fs = require('fs');
 const readLastLines = require('read-last-lines');
 const axios = require('axios');
@@ -19,15 +20,13 @@ const radiantBackendEndpoints = {
 const streamTracker = {
 
 };
-
 let watcher;
 
 module.exports.watcher = (ouPath, args) => {
-    console.log(`watcher started for : ${ouPath}`);
+    // console.log(`watcher started for : ${ouPath}`);
 
     watcher = chokidar.watch(ouPath);
     const authToken = args.token;
-    // call watcher close?
     watcher.on('add', function (path) {
         //check file
         streamTracker[path] = {
@@ -56,7 +55,7 @@ module.exports.end = (streamPath) => {
             if(err){
                 console.log('-=*[ERROR: no such file or directory ]*=-');
             }
-            console.log('-=*[ Cleanup remaining files ]*=-');
+            // console.log('-=*[ Cleanup remaining files ]*=-');
             files.forEach(file => {
                 const fileC = file.split('.')[1];
                 if(fileC === 'm3u8') {
@@ -112,11 +111,13 @@ const checkFile = function (info){
     setTimeout((args) => {
         fs.stat(args[0].info.path, (err, fileInfo) => {
             if(err === null) {
-                if(fileInfo.size === 0) {
+                if(fileInfo.size <= 50000) {
                     // console.log(`-=*[ checking file: ${info.path} with size: ${fileInfo.size}: checking again in 1.5 sec ]*=-`);
                     if(streamTracker[info.path].retry <= 3){
                         streamTracker[info.path].retry++;
                         checkFile(info);
+                    } else {
+                        uploadFile(info);
                     }
                 } else {
                     const ext = info.path.replace(/^.*[\\\/]/, '').split('.')[1];
@@ -128,7 +129,7 @@ const checkFile = function (info){
                 }
             }
         });
-    }, 1500, [{
+    }, 1000, [{
         info
     }]);
 };
@@ -160,13 +161,16 @@ const uploadFile = function (info){
                 if(ext === 'm3u8' && !streamTracker[info.path].m3u8){
                     streamTracker[info.path].m3u8 = true;
                     console.log(`-=*[ CREATING VIDEO STREAM ]*=-`);
-                    console.log(`-=*[ conversationTopicId = ${info.conversationTopicId} ]*=-`);
-                    console.log(`-=*[ auth token = ${info.authToken} ]*=-`);
+                    // console.log(`-=*[ conversationTopicId = ${info.conversationTopicId} ]*=-`);
+                    // console.log(`-=*[ auth token = ${info.authToken} ]*=-`);
                     createVideoStream(info.conversationTopicId, info.authToken)
                         .then((vidData) => updateVideoStream(vidData, data.Key, mainPath, info.authToken)
                             .then((res) => {
-                                console.log(`-=*[ StreamID = : ${res.liveStream.updateStream.id} ]*=-`);
-                                console.log(`-=*[ Stream downloadUrl : ${res.liveStream.updateStream.downloadUrl.url} ]*=-`);
+                                console.log(`-=*[ StreamID = : ${res.videoStreamData.liveStream.updateStream.id} ]*=-`);
+                                console.log(`-=*[ Stream downloadUrl : ${res.videoStreamData.liveStream.updateStream.downloadUrl.url} ]*=-`);
+                                return createThumbnail(mainPath, `${data.Key.split('-')[0]}`).then((thumbnailInfo) => {
+                                    return uploadThumbnail(thumbnailInfo.thumbnailPath, thumbnailInfo.videoPath, data.Key.split('-')[0],  info.authToken, res.vidData.conversationTopic.createConversationTopicVideo.video.id);
+                                });
                             })).catch((err => {
                         console.log(err);
                     }));
@@ -174,7 +178,6 @@ const uploadFile = function (info){
             } catch (e) {
                 console.log(`ERROR: ${e.message} not too big of a deal :D`);
             }
-
 
             if(ext === 'ts'){
                 // upload m3u8 to keep it updated
@@ -187,15 +190,19 @@ const uploadFile = function (info){
                 // console.log(`-=*[ UPDATE: uploading file: ${mainPath}/${m3u8}-i.m3u8 ]*=-`);
                 // delete ts file
                 // console.log(`deleting file => ${info.path}`);
-                fs.stat(info.path, (err) => {
-                    if(err === null) {
-                        fs.unlink(info.path, (err, data) => {
-                            if(err){
-                                console.log(err);
-                            }
-                        });
-                    }
-                });
+                if(info.path === `${mainPath}/${m3u8}-i1.ts`){
+                    // dont delete
+                } else {
+                    fs.stat(info.path, (err) => {
+                        if(err === null) {
+                            fs.unlink(info.path, (err, data) => {
+                                if(err){
+                                    console.log(err);
+                                }
+                            });
+                        }
+                    });
+                }
             } else if(ext === 'm3u8'){
                 const m3u8 = data.Key.split('-')[0];
                 checkM3U8(`${mainPath}/${m3u8}-i.m3u8`);
@@ -239,7 +246,6 @@ const query = gql`
  * @returns {Promise<T | never>}
  */
 const createVideoStream = function(conversationTopicId, authToken) {
-    console.log(`-=*[ auth token = ${authToken} ]*=-`);
     const options = {
         headers: {
             Accept: "application/json",
@@ -257,9 +263,9 @@ const createVideoStream = function(conversationTopicId, authToken) {
         variables,
     }, options).then((results) => {
         console.log('-=*[ CREATED VIDEO STREAM ]*=-');
-        console.log(`-=*[ Conversation Topic Id = ${conversationTopicId} ]*=-`);
-        console.log(`-=*[ Video Id = ${results.data.data.conversationTopic.createConversationTopicVideo.video.id} ]*=-`);
-        console.log(`-=*[ Video Stream Id = ${results.data.data.conversationTopic.createConversationTopicVideo.videoHLSStreamUpload.id} ]*=-`);
+        // console.log(`-=*[ Conversation Topic Id = ${conversationTopicId} ]*=-`);
+        // console.log(`-=*[ Video Id = ${results.data.data.conversationTopic.createConversationTopicVideo.video.id} ]*=-`);
+        // console.log(`-=*[ Video Stream Id = ${results.data.data.conversationTopic.createConversationTopicVideo.videoHLSStreamUpload.id} ]*=-`);
         return results.data.data;
     }).catch((err) => {
        console.log('ERROR -- created Video Stream');
@@ -295,7 +301,7 @@ const updateVideoStream = function(vidData, key, mainPath, authToken) {
     const options = {
         headers: {
             Accept: "application/json",
-            subauth: `${authToken}`,
+            subauth: `Bearer ${authToken}`,
             "Content-Type": "application/json"
         }
     };
@@ -306,17 +312,167 @@ const updateVideoStream = function(vidData, key, mainPath, authToken) {
     let endpoint = radiantBackendEndpoints[process.env.ENV];
 
     console.log(`-=*[ UPDATING VIDEO STREAM ]*=-`);
-    console.log(`-=*[ key = ${key} ]*=-`);
+    // console.log(`-=*[ key = ${key} ]*=-`);
     // i have thumbnail upload url here
     return axios.post(endpoint, {
         query: print(videoStreamQuery),
         variables,
     }, options).then((results) => {
         console.log('-=*[ UPDATING VIDEO STREAM ]*=-');
-        console.log(`-=*[ ${results.data.data.liveStream.updateStream.downloadUrl.url} ]*=-`);
-        return results.data.data;
+        console.log(`-=*[ m3u8 : ${results.data.data.liveStream.updateStream.downloadUrl.url} ]*=-`);
+        return {
+            vidData,
+            videoStreamData: results.data.data
+        };
     }).catch((err) => {
         console.log('ERROR -- Updated Video Stream');
         console.log(err);
+    });
+};
+
+const updateVideoQuery = gql`
+    mutation updateVideo($id: ID!, $thumbnail: String!){
+        updateVideo(input:{
+            id:$id
+            thumbnail:$thumbnail
+        }){
+            id
+            thumbnailUrl
+        }
+    }
+`;
+
+/**
+ * updateVideo
+ * @param videoId
+ * @param thumbnailUrl
+ * @param authToken
+ * @returns {Promise<T | never>}
+ */
+const updateVideo = function(videoId, thumbnailUrl, authToken){
+    const options = {
+        headers: {
+            Accept: "application/json",
+            subauth: `Bearer ${authToken}`,
+            "Content-Type": "application/json"
+        }
+    };
+    const variables = {
+        id: videoId,
+        thumbnail: thumbnailUrl,
+    };
+    let endpoint = radiantBackendEndpoints[process.env.ENV];
+
+    console.log(`-=*[ UPDATING VIDEO ]*=-`);
+    console.log(`-=*[ VideoId = ${videoId} ]*=-`);
+    // i have thumbnail upload url here
+    return axios.post(endpoint, {
+        query: print(updateVideoQuery),
+        variables,
+    }, options).then((results) => {
+        console.log('-=*[ UPDATED VIDEO ]*=-');
+        console.log(`-=*[ Video Id : ${results.data.data.updateVideo.id} ]*=-`);
+        console.log(`-=*[ Thumbnail Url: ${thumbnailUrl} ]*=-`);
+        return results;
+    }).catch((err) => {
+        console.log('ERROR -- Updated Video');
+        console.log(err);
+    });
+};
+
+/**
+ * uploadThumbnail
+ * @param thumb
+ * @param videoPath
+ * @param fileKey
+ * @param authToken
+ * @param videoId
+ */
+const uploadThumbnail = function(thumb, videoPath, fileKey, authToken, videoId){
+    const params = {
+        Bucket: process.env.S3_BUCKET,
+        Key: fileKey,
+        Body: fs.createReadStream(thumb),
+        ACL: 'public-read',
+        ContentType: 'image/png',
+    };
+    // upload thumbnail
+    AWS.getS3().upload(params, (err, data) => {
+        if(err){
+            console.log(err);
+        } else {
+            console.log(data);
+            // update thumbnail on video record
+            updateVideo(videoId, data.Location, authToken).then((data) => {
+                console.log(`VIDEO UPDATED SUCCESS => ${data.data.data.updateVideo.id}`);
+                // delete thumbnail
+                fs.unlink(thumb, (err) => {
+                    if(err){
+                        console.log(`Error Deleting thumbnail for ${fileKey}: ${err}`);
+                    } else {
+                        // console.log(`Deleted File: ${fileKey}`);
+                    }
+                });
+                // delete thumbnail video file reference
+                fs.unlink(videoPath, (err) => {
+                    if(err){
+                        console.log(`Error Deleting video reference for thumbnail: ${videoPath}: ${err}`);
+                    } else {
+                        // console.log(`Deleted File: ${videoPath}`);
+                    }
+                });
+            }).catch((err) => {
+                console.log(`UPDATE VIDEO ERROR: ${err}`);
+            });
+        }
+    });
+};
+/**
+ * createThumbnail
+ * @param mainPath
+ * @param fileKey
+ * @returns {Promise}
+ */
+const createThumbnail = function(mainPath, fileKey) {
+    return new Promise((resolve, reject) => {
+        console.log(`-=*[ Thumbnail Creation ]*=-  ${fileKey}`);
+        const thumbnailPath = `media/thumbnails/${fileKey}.png`;
+        const videoPath = `${mainPath}/${fileKey}-i1.ts`;
+        setTimeout(() =>{
+        fs.stat(videoPath, (err, data) => {
+           if(err === null){
+               const argv = [
+                   '-i',
+                   videoPath,
+                   '-ss',
+                   '00:00:01',
+                   '-c:v',
+                   'mjpeg',
+                   '-f',
+                   'mjpeg',
+                   '-vframes',
+                   '1',
+                   thumbnailPath,
+               ];
+               const ffmpegSpawn = spawn(process.env.FFMPEG_PATH, argv);
+               ffmpegSpawn.on('error', (e) => {
+                   console.log(`Error Creating Thumbnail: ${e}`);
+                   reject(e);
+               });
+               ffmpegSpawn.stdout.on('data', (d) => {
+                   console.log(`Thumbnail: ${d}`);
+               });
+               ffmpegSpawn.stderr.on('data', (d) => {
+                   console.log(`Thumbnail: ${d}`);
+               });
+               ffmpegSpawn.on('close', (c) => {
+                   console.log(`Thumbnail Close: ${c}`);
+                   resolve({ thumbnailPath, videoPath });
+               });
+           } else {
+               console.log(`No Video File: ${err}`);
+           }
+        });
+        }, 1000);
     });
 };
